refactor(desktop): split select and accordion primitives
